Return embedded resource if there is one
Making Resource#get return embedded resources before fetching them from
the server, makes it possible for the server to decided/change which
resources should be preloaded. The user of ShafClient can then be
happily unaware of server changes where embedded resources are added
to/remove from a certain response.
Currently there is one issue with this approach. We loose the
possibility to deserialize the embedded resource into anything but
ShafClient::Resource, since we don't know the content type.