Add support for multiple hypertext cache strategies
The Hypertext Cache Pattern allows the server to embed reources in
addition to linking to them. This means that if we call #get(rel) we can
directly return the embedded resource (if present) instead of fetching
it from the server. However, this can cause problems if the embedded
resource should not be interpreted as application/hal+json (i.e. plain
HAL without any extensions).
This commit adds three possibiltiest for handling this.
1. Ignore embedded resources and refetch them through link.
2. Return the embedded resource (as regular HAL resource).
3. Perform a HEAD request and return the resource with correct headers.